feat. FavoriteGroupList component 구현
